working to separate filters from results
